+++ schoolsite.client/src/app/components/gallery-image/gallery-image-table/gallery-image-table.component.ts
@@ -1,7 +1,9 @@
-import { Component } from '@angular/core';
+import { Component, ElementRef, ViewChild } from '@angular/core';
 import { GalleryImage } from '../../../../entities/galleryImage';
 import { GalleryImageService } from '../../../services/gallery-image.service';
 import { ActivatedRoute, Router } from '@angular/router';
+
+declare var bootstrap: any;
 
 @Component({
   selector: 'gallery-image-table',
@@ -11,6 +13,10 @@ import { ActivatedRoute, Router } from '@angular/router';
 export class GalleryImageTableComponent {
   galleryImages: GalleryImage[] = [];
   galleryId!: number;
+  selectedGalleryImageId: number | null = null;
+  modalInstance: any;
+
+  @ViewChild('exampleModal') exampleModal!: ElementRef;
 
   constructor(
     private galleryImageService: GalleryImageService,
@@ -28,6 +34,13 @@ export class GalleryImageTableComponent {
     });
   }
 
+  ngAfterViewInit() {
+    // Initialize the modal instance
+    if (this.exampleModal) {
+      this.modalInstance = new bootstrap.Modal(this.exampleModal.nativeElement);
+    }
+  }
+
   loadGalleryImages(galleryId: number): void {
     this.galleryImageService.getGalleryImagesByGalleryId(galleryId).subscribe({
       next: (data: GalleryImage[]) => {
@@ -40,9 +53,19 @@ export class GalleryImageTableComponent {
   }
 
   deleteGalleryImage(id: number): void {
+    if (id === null) {
+      console.error('No Id selected for deletion');
+      return;
+    }
+
     this.galleryImageService.deleteGalleryImage(id).subscribe({
       next: () => {
         this.galleryImages = this.galleryImages.filter((gi) => gi.id != id);
+
+        //Close the modal
+        if (this.modalInstance) {
+          this.modalInstance.hide();
+        }
       },
       error: (err) => {
         console.error('Error for deleting gallery image', err);
