+++ schoolsite.client/src/app/components/team-member/team-member-table/team-member-table.component.ts
@@ -1,18 +1,26 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, ElementRef, ViewChild } from '@angular/core';
 import { TeamMember } from '../../../../entities/teamMember';
 import { TeamMemberService } from '../../../services/team-member.service';
 import { Router } from '@angular/router';
 
+declare var bootstrap: any;
+
 @Component({
   selector: 'team-member-table',
   templateUrl: './team-member-table.component.html',
-  styleUrl: './team-member-table.component.css'
+  styleUrl: './team-member-table.component.css',
 })
 export class TeamMemberTableComponent {
-
   teamMembers: TeamMember[] = [];
+  selectedTeamMemberId: number | null = null;
+  modalInstance: any;
 
-  constructor(private teamMemberService: TeamMemberService, private router: Router){}
+  @ViewChild('exampleModal') exampleModal!: ElementRef;
+
+  constructor(
+    private teamMemberService: TeamMemberService,
+    private router: Router
+  ) {}
 
   ngOnInit() {
     this.teamMemberService.getTeamMembers().subscribe((data: TeamMember[]) => {
@@ -21,19 +29,35 @@ export class TeamMemberTableComponent {
     });
   }
 
-  deleteTeamMember(id: number) : void {
+  ngAfterViewInit() {
+    // Initialize the modal instance
+    if (this.exampleModal) {
+      this.modalInstance = new bootstrap.Modal(this.exampleModal.nativeElement);
+    }
+  }
+
+  deleteTeamMember(id: number): void {
+    if (id === null) {
+      console.error('No Id selected for deletion');
+      return;
+    }
+
     this.teamMemberService.deleteTeamMember(id).subscribe({
       next: () => {
-        this.teamMembers = this.teamMembers.filter(tm => tm.id != id);
+        this.teamMembers = this.teamMembers.filter((tm) => tm.id != id);
+
+        //Close the modal
+        if (this.modalInstance) {
+          this.modalInstance.hide();
+        }
       },
       error: (err) => {
         console.error('Error for deleting team member', err);
-      }
+      },
     });
   }
 
-  editTeamMember(id: number) : void {
+  editTeamMember(id: number): void {
     this.router.navigate(['/team-members/edit', id]);
   }
-
 }
